inside components folder, separated the files into CardForm, Navbar, Images and AppStyles
