+++ src/components/NavBar/NavBar.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Logo from '../Images/pepsi.png';
+
+const NavBar = () => {
+	return (
+		<img src={Logo} width="10%" height="10%" alt="Logo" />
+		)
+}
+
+export default NavBar;
